--- src/main/java/TaskParser.java
@@ -1,4 +1,0 @@
-@FunctionalInterface
-public interface TaskParser {
-    Task get() throws DukeException;
-}
