+++ src/main/java/duke/command/IndexedCommand.java
@@ -1,3 +1,5 @@
+package duke.command;
+
 public abstract class IndexedCommand extends Command {
     protected int index;
