+++ src/main/java/duke/command/ListCommand.java
@@ -1,3 +1,5 @@
+package duke.command;
+
 public class ListCommand extends Command {
     ListCommand() {
         super.isExit = false;
